# perf: Comprehensive authentication system optimization