… AUTH_OPTIMIZATION_README.md
… docs/AUTH_OPTIMIZATION_TESTING_GUIDE.md
… docs/AUTH_PERFORMANCE_OPTIMIZATION.md
… supabase/migrations/20250103_auth_performance_indexes.sql